factored out chart dataset eval to parent component
