+++ packages/runtime/src/widgets/Charts/BarChart.tsx
@@ -1,12 +1,7 @@
 import { Bar } from "react-chartjs-2";
 import type { ChartOptions } from "chart.js";
-import {
-  useRegisterBindings,
-  evaluate,
-  useScreenContext,
-  type ScreenContextDefinition,
-} from "@ensembleui/react-framework";
-import { useMemo, useState } from "react";
+import { useRegisterBindings } from "@ensembleui/react-framework";
+import { useState } from "react";
 import type { ChartDataSets, ChartProps } from "..";
 
 const options: ChartOptions<"bar"> = {
@@ -38,21 +33,11 @@ export const BarChart: React.FC<ChartProps> = (props) => {
 
   const [title, setTitle] = useState(config?.title);
   const [labels, setLabels] = useState<string[]>(config?.data?.labels || []);
-  const context = useScreenContext();
-
-  const evaluatedDatasets = useMemo(
-    () => evaluate(context as ScreenContextDefinition, config?.data?.datasets),
-    [config?.data?.datasets, context],
-  );
 
   const { values } = useRegisterBindings({ labels, title }, id, {
     setLabels,
     setTitle,
   });
-
-  console.log("BarChart", evaluatedDatasets);
-
-  if (!evaluatedDatasets) return null;
 
   return (
     <div
@@ -64,7 +49,7 @@ export const BarChart: React.FC<ChartProps> = (props) => {
       <Bar
         data={{
           labels: values?.labels,
-          datasets: evaluatedDatasets as ChartDataSets[],
+          datasets: config?.data?.datasets as ChartDataSets[],
         }}
         options={{
           ...options,
