+++ packages/runtime/src/widgets/Charts/index.tsx
@@ -11,8 +11,13 @@ import {
   LineElement,
   type ChartOptions,
 } from "chart.js";
-import React, { cloneElement } from "react";
-import type { Expression } from "@ensembleui/react-framework";
+import React, { cloneElement, useMemo } from "react";
+import {
+  evaluate,
+  type ScreenContextDefinition,
+  type Expression,
+  useScreenContext,
+} from "@ensembleui/react-framework";
 import { WidgetRegistry } from "../../registry";
 import type { EnsembleWidgetProps } from "../../util/types";
 import { BarChart } from "./BarChart";
@@ -65,18 +70,34 @@ const tabsConfig = {
 };
 
 export const Chart: React.FC<ChartProps> = (props) => {
+  const context = useScreenContext();
+
   // eslint-disable-next-line @typescript-eslint/no-implied-eval, no-new-func
   const config = new Function(
     `return ${props.config?.toString() || "{}"}`,
   )() as ChartConfigs;
 
-  if (!config.type) {
-    return <b>Chart type missing</b>;
-  }
+  const evaluatedDatasets = useMemo(
+    () =>
+      evaluate(
+        context as ScreenContextDefinition,
+        JSON.stringify(config?.data?.datasets),
+      ),
+    [config?.data?.datasets, context],
+  );
+
+  if (!config.type) return <b>Chart type missing</b>;
+  if (!evaluatedDatasets) return null;
 
   return cloneElement(tabsConfig[config.type], {
     ...props,
-    config,
+    config: {
+      ...config,
+      data: {
+        ...config.data,
+        datasets: evaluatedDatasets,
+      },
+    },
   });
 };
